Метод пересоздания размера вынесен в класс сервиса размеров
